requirements.txt
requests
beautifulsoup4
pandas
apache-airflow-providers-mysql
mysql-connector-python
matplotlib